\name{greenampt}
\alias{greenampt}
\title{greenampt: Adjustment to the Green-Ampt model of an infiltration test}
\usage{
greenampt(data)
}
\description{
Adjustment to the Green-Ampt model of an infiltration test. Green-Ampt equation and plot of observed and simulated infiltration.
}
\arguments{
data: A dataframe containing a point measurement by infiltrometer: time in min (time), accumulated depth in cm (depth)
}
\examples{
greenampt(data)
}
